moved list items to BlockScrollList component
